--- app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/menu/income/DeleteItemClickListener.kt
@@ -1,7 +1,0 @@
-package com.romandevyatov.bestfinance.ui.adapters.menu.income
-
-interface DeleteItemClickListener<T> {
-
-    fun deleteItem(item: T)
-
-}
