+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/transactions_deprecated/expense/ChildExpenseSubGroupsAdapter.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.adapters.menu.expense;
+package com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.expense;
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
