+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/transactions_deprecated/expense/ParentExpenseGroupAdapter.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.adapters.menu.expense
+package com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.expense
 
 
 import android.view.LayoutInflater
