+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/transactions_deprecated/income/ArchiveItemBySwipe.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.adapters.menu.income
+package com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.income
 
 import com.romandevyatov.bestfinance.db.entities.IncomeSubGroup
