+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/transactions_deprecated/income/DeleteItemClickListener.kt
@@ -1,0 +1,7 @@
+package com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.income
+
+interface DeleteItemClickListener<T> {
+
+    fun deleteItem(item: T)
+
+}
