+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/deprecated/IncomeFragment.kt
@@ -18,8 +18,8 @@ import com.romandevyatov.bestfinance.db.entities.IncomeHistory
 import com.romandevyatov.bestfinance.db.entities.IncomeSubGroup
 import com.romandevyatov.bestfinance.db.entities.relations.IncomeGroupWithIncomeSubGroupsIncludingIncomeHistories
 import com.romandevyatov.bestfinance.db.entities.relations.IncomeSubGroupWithIncomeHistories
-import com.romandevyatov.bestfinance.ui.adapters.menu.income.ArchiveItemBySwipe
-import com.romandevyatov.bestfinance.ui.adapters.menu.income.ParentIncomeGroupAdapter
+import com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.income.ArchiveItemBySwipe
+import com.romandevyatov.bestfinance.ui.adapters.transactions_deprecated.income.ParentIncomeGroupAdapter
 import com.romandevyatov.bestfinance.ui.adapters.clicklisteners.AddItemClickListener
 import com.romandevyatov.bestfinance.viewmodels.foreachmodel.IncomeGroupViewModel
 import com.romandevyatov.bestfinance.viewmodels.foreachmodel.IncomeHistoryViewModel
