+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/history/ExpenseHistoryFragment.kt
@@ -1,6 +1,5 @@
 package com.romandevyatov.bestfinance.ui.fragments.history
 
-
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -12,7 +11,6 @@ import com.romandevyatov.bestfinance.databinding.FragmentExpenseHistoryBinding
 import com.romandevyatov.bestfinance.ui.adapters.history.ExpenseHistoryAdapter
 import com.romandevyatov.bestfinance.viewmodels.foreachmodel.ExpenseHistoryViewModel
 import dagger.hilt.android.AndroidEntryPoint
-
 
 @AndroidEntryPoint
 class ExpenseHistoryFragment : Fragment() {
