+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/history/HistoryFragment.kt
@@ -10,7 +10,6 @@ import com.romandevyatov.bestfinance.databinding.FragmentHistoryBinding
 import com.romandevyatov.bestfinance.ui.adapters.history.ViewPagerAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
-
 @AndroidEntryPoint
 class HistoryFragment : Fragment() {
