+++ app/src/main/java/com/romandevyatov/bestfinance/utils/Constants.kt
@@ -9,6 +9,8 @@ object Constants {
     const val ADD_NEW_SUB_INCOME_GROUP = "Add new sub income group"
     const val ADD_NEW_INCOME_GROUP = "Add new income group"
     const val DATABASE_NAME = "bestfinance_database"
+    const val WALLET_TO = "Wallet to"
+    const val WALLET_FROM = "Wallet from"
 
     const val PARENT = 0
     const val CHILD = 1
@@ -18,6 +20,5 @@ object Constants {
 
     const val FIELD_IS_EMPTY_MESSAGE = "Please enter some text!"
     const val ADD_NEW_WALLET_STRING: String = "Add new wallet"
-
 
 }
